Rename featuregate top level directory to featuregates
Plural sounds better.
