+++ cmd/cli/plugin/feature/go.mod
@@ -7,7 +7,7 @@ replace (
 	github.com/vmware-tanzu/tanzu-framework/apis/config => ./../../../../apis/config
 	github.com/vmware-tanzu/tanzu-framework/apis/run => ./../../../../apis/run
 	github.com/vmware-tanzu/tanzu-framework/cli/runtime => ./../../../../cli/runtime
-	github.com/vmware-tanzu/tanzu-framework/featuregate/client => ./../../../../featuregate/client
+	github.com/vmware-tanzu/tanzu-framework/featuregates/client => ./../../../../featuregates/client
 	sigs.k8s.io/cluster-api => sigs.k8s.io/cluster-api v1.1.5
 )
 
@@ -17,7 +17,7 @@ require (
 	github.com/vmware-tanzu/tanzu-framework/apis/cli v0.0.0-00010101000000-000000000000
 	github.com/vmware-tanzu/tanzu-framework/apis/config v0.0.0-20220820022318-db127cf83a25
 	github.com/vmware-tanzu/tanzu-framework/cli/runtime v0.0.0-20220901171806-254f018f3ce4
-	github.com/vmware-tanzu/tanzu-framework/featuregate/client v0.0.0-00010101000000-000000000000
+	github.com/vmware-tanzu/tanzu-framework/featuregates/client v0.0.0-00010101000000-000000000000
 	k8s.io/apimachinery v0.23.5
 	k8s.io/client-go v0.23.5
 	sigs.k8s.io/controller-runtime v0.11.2
