+++ cmd/cli/plugin/feature/list.go
@@ -13,7 +13,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/vmware-tanzu/tanzu-framework/cli/runtime/component"
-	"github.com/vmware-tanzu/tanzu-framework/featuregate/client/pkg/featuregateclient"
+	"github.com/vmware-tanzu/tanzu-framework/featuregates/client/pkg/featuregateclient"
 )
 
 var featuregate, outputFormat string
