+++ cmd/cli/plugin/feature/list_test.go
@@ -11,8 +11,8 @@ import (
 	crclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	configv1alpha1 "github.com/vmware-tanzu/tanzu-framework/apis/config/v1alpha1"
-	"github.com/vmware-tanzu/tanzu-framework/featuregate/client/pkg/featuregateclient"
-	"github.com/vmware-tanzu/tanzu-framework/featuregate/client/pkg/featuregateclient/fake"
+	"github.com/vmware-tanzu/tanzu-framework/featuregates/client/pkg/featuregateclient"
+	"github.com/vmware-tanzu/tanzu-framework/featuregates/client/pkg/featuregateclient/fake"
 )
 
 func TestJoinFeatures(t *testing.T) {
