+++ featuregates/client/pkg/featuregateclient/client_test.go
@@ -12,7 +12,7 @@ import (
 	crclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	configv1alpha1 "github.com/vmware-tanzu/tanzu-framework/apis/config/v1alpha1"
-	"github.com/vmware-tanzu/tanzu-framework/featuregate/client/pkg/featuregateclient/fake"
+	"github.com/vmware-tanzu/tanzu-framework/featuregates/client/pkg/featuregateclient/fake"
 )
 
 const contextTimeout = 30 * time.Second
