+++ featuregates/controller/pkg/featuregate/doc.go
@@ -1,5 +1,5 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Package controllers has the controller for feature gates.
+// Package featuregate has the controller for feature gates.
 package featuregate
